@@ -1,36 +1,12 @@
-import React, { Component } from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import "./App.css";
-import Navbar from "./components/Navbar";
-import Home from "./components/Home";
-import Products from "./components/Products";
-import ProductAdmin from "./components/ProductAdmin";
-import Footer from "./components/Footer";
-import { library } from "@fortawesome/fontawesome-svg-core";
-import { faEdit } from "@fortawesome/free-solid-svg-icons";
-import axios from "axios";
-const config = require("./config.json");
-library.add(faEdit);
-axios.defaults.headers.common["x-api-key"] = config.api.key;
+import Bulletin from "./Component/Bulletin";
 
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <Router>
-          <div>
-            <Navbar />
-            <Switch>
-              <Route exact path="/" component={Home} />
-              <Route exact path="/products" component={Products} />
-              <Route exact path="/admin" component={ProductAdmin} />
-            </Switch>
-            <Footer />
-          </div>
-        </Router>
-      </div>
-    );
-  }
+function App() {
+  return (
+    <div className="App">
+      <Bulletin />
+    </div>
+  );
 }
 
 export default App;
